% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Population_Palette.R
\name{areColors}
\alias{areColors}
\title{Internal - check if characters in a vector are valid colors}
\usage{
areColors(x)
}
\arguments{
\item{x}{a character vector of color names}
}
\value{
a logical vector
}
\description{
Internal - check if characters in a vector are valid colors
}
