% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Population_Palette.R
\name{changeLevel}
\alias{changeLevel}
\title{Replaces a CYTdata's clustering slot with its metaclustering slot, or vice-versa}
\usage{
changeLevel(
  CYTdata,
  direction = c("clustersToMetaclusters", "metaclustersToClusters"),
  checkOverwrite = TRUE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{direction}{a string specifying the direction of the change, "clustersToMetaclusters" or "metaclustersToClusters"}

\item{checkOverwrite}{a boolean value indicating whether to check if a clustering/metaclustering has already been performed on the CYTdata object.}
}
\value{
a CYTdata object
}
\description{
This functions copy-pastes the clustering slot into the metaclustering slot or the metaclustering slot into the clustering slot. It does not empty the copy-pasted slot.
}
