% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Population_Palette.R
\name{changePalettes}
\alias{changePalettes}
\title{Change color palette associated with clusters/metaclusters}
\usage{
changePalettes(
  CYTdata,
  level = c("clusters", "metaclusters"),
  autoColor = TRUE,
  population = NULL,
  autoColorType = c("rainbow", "blank"),
  homemadePalette = NULL
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{level}{a character value indicating whether to change palette on clusters or metaclusters. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{autoColor}{logical. Whether to generate a color palette automatically (TRUE), or provide a custom one (FALSE)}

\item{population}{a character vector containing the identifiers of the population(s) (clusters or metaclusters) for which to change the color. If NULL, (default) all clusters/metaclusters are used}

\item{autoColorType}{character string specifying the type of color palette to generate if autoColor is set to TRUE. Possible values are "rainbow" (default) and "blank", which will set all colors to white (#FFFFFF)}

\item{homemadePalette}{a named character vector containing the colors (hexadecimal code, or name) to be associated with each cluster/metacluster in population. Necessary if autoColor = FALSE.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to change the color palette associated with a specific set of clusters or metaclusters
}
