% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering_QC.R
\name{check_HierarchyMetaclusters}
\alias{check_HierarchyMetaclusters}
\title{Check that no cluster is divided between two metaclusters}
\usage{
check_HierarchyMetaclusters(CYTdata, output = c("message", "stop"))
}
\arguments{
\item{CYTdata}{a CYTdata object with clustering and metaclustering performed}

\item{output}{a character vector indicating whether to display a message or outright stop the function if a cluster is divided between two metaclusters.}
}
\description{
Check that no cluster is divided between two metaclusters
}
