% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_FCS_Markers_CYTdata.R
\name{checkorderMarkers}
\alias{checkorderMarkers}
\title{Checks marker name conformity within a CYTdata object}
\usage{
checkorderMarkers(CYTdata, markers, order = TRUE, checkDuplicates = TRUE)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a vector object listing marker names}

\item{order}{a boolean specifying whether to reorder the markers according to colnames of the CYTdata's matrix expression}

\item{checkDuplicates}{a boolean specifying whether to check for duplicates in the markers list}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to check for marker name conformity
It will check whether a marker list is identical to the colnames of the CYTdata's marker expression matrix
If order = T, it will also order the marker according to colnames.
}
