% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Population_Palette.R
\name{checkorderPopulation}
\alias{checkorderPopulation}
\title{Check and order population names according to clusters/metaclusters}
\usage{
checkorderPopulation(
  CYTdata,
  population,
  level,
  order = TRUE,
  checkDuplicates = TRUE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{population}{a character vector containing the identifiers of the population(s) (clusters or metaclusters) to check and potentially order}

\item{level}{a character value indicating whether to compare population to the clusters or the metaclusters. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{order}{logical. If TRUE, population will be ordered according to the order of the levels of the clusters/metaclusters}

\item{checkDuplicates}{logical. If TRUE, will check for duplicates inside the given population character vector.}
}
\value{
a character vector
}
\description{
This functions checks that names in a character vector are all present in a CYTdata's clustering or metaclustering object
If order = TRUE, it also orders the character vector according to the order of the levels of the CYTdata's clusters/metaclusters.
If checkDuplicates = TRUE, it also checks for duplicates in the character vector 
The function is useful, for example, to check that no mistakes were made when writing a list of cluster names.
}
