% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Metadata_Samples.R
\name{checkorderSamples}
\alias{checkorderSamples}
\title{Checks sample name conformity within a CYTdata object}
\usage{
checkorderSamples(CYTdata, samples, order = TRUE, checkDuplicates = TRUE)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{samples}{a vector object listing sample names}

\item{order}{a boolean specifying whether to reorder the samples according to the levels of the CYTdata's samples factor}

\item{checkDuplicates}{a boolean specifying whether to check for duplicates in the given sample list.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to check for sample name conformity
It will check whether a sample list is identical to levels of the CYTdata's samples factor
If order = T, it will also order the samples according to the CYTdata's samples factor
}
