% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{Clustering-class}
\alias{Clustering-class}
\alias{Clustering}
\title{Clustering class definition}
\description{
The Clustering object is a S4 object containing all the data concerning the Clustering step performed on a 'CYTdata' object
}
\section{Slots}{

\describe{
\item{\code{clusters}}{a vector containing the identifiers of cell clusters as a factor}

\item{\code{cellcount}}{a data.frame containing the number of cells associated to each cluster for each sample}

\item{\code{abundance}}{a data.frame containing the percentage of cells associated to each cluster for each sample}

\item{\code{palette}}{TO DO
#' @slot optional_parameters TO CHECK a vector containing the parameters used for the identification of the cell clusters (especially the argument of 'run.Clustering' function : seed, markers, architecture, type, etc.)}

\item{\code{optional_plots}}{TO DO}
}}

