% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Clustering.R
\name{compute.abundance}
\alias{compute.abundance}
\title{Internal - Computes the relative abundances for each cell population}
\usage{
\method{compute}{abundance}(cellcount)
}
\arguments{
\item{cellcount}{a data.frame providing the numbers of cells associated to each population for each sample}
}
\value{
a data.frame containing the abundance of cells to each population for each sample
}
\description{
This function is used internally to computes the abundance of each population for each sample.
}
