% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Clustering.R
\name{compute.cellcount}
\alias{compute.cellcount}
\title{Internal - Computes the number of cells for each population}
\usage{
compute.cellcount(population, samples)
}
\arguments{
\item{population}{a character vector providing the population to count the cells belonging to}

\item{samples}{a character vector providing for each cell the associated biological sample}
}
\value{
a dataframe containing the numbers of cells associated for each population for each sample (rownames = population / colnames = samples)
}
\description{
This function is used internally to computes the number of cells for each population.
}
