% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_DimReduction_QC.R
\name{computeDimRed_CorrelationQC}
\alias{computeDimRed_CorrelationQC}
\title{QC for dimensionality reduction: correlation between pairwise distances in the high-dimensional space and in the reduced space}
\usage{
computeDimRed_CorrelationQC(
  CYTdata,
  subsetSize = 1000,
  method = c("pearson", "kendall", "spearman"),
  markers = NULL
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{subsetSize}{a numeric being the sample size used for computation. Defaults to 1000. Higher values will lead to memory saturation.}

\item{method}{a character being the method use to compute correlation. Possible values are : "pearson", "kendall", "spearman". Default value is "pearson"}

\item{markers}{a character vector of the markers to consider for the knn. Defaults to the markers used to compute the dimensionnality reduction (recommended).}
}
\value{
a list containing correlation coefficient and boxplot of pairwise distances
}
\description{
Compares distances between cells in the original, high-dimensional space and in the reduced space. This shows the preservation of the global structure in the dimension reduced space. 
Theoretically, a "perfect" dimension reduction should give a linear relationships between pairwise distances in both data spaces.
}
