% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_DimReduction_QC.R
\name{computeDimRed_KNCQC}
\alias{computeDimRed_KNCQC}
\title{QC for dimensionality reduction : Computes proportion of nearest clusters preservation in reduced dimension in comparison to high dimension}
\usage{
computeDimRed_KNCQC(CYTdata, knc = 5, subsetSize = NULL, markers = NULL)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{knc}{a numeric being the number of nearest classes to compute. Defaults to 5.}

\item{subsetSize}{a numeric being the sample size used for computation.}

\item{markers}{a character vector of the markers to consider for the knn. Defaults to the markers used to compute the dimensionnality reduction (recommended).}
}
\value{
a numeric proportion
}
\description{
Computes the fraction of k-nearest class (clusters) means in the original data that are preserved as k-nearest class means in the reduced data. 
Computes the class mean only and averages across all classes. KNC quantifies preservation of the mesoscopic structure.
}
