% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_DimReduction_QC.R
\name{computeDimRed_KNNQC}
\alias{computeDimRed_KNNQC}
\title{QC for dimensionality reduction : Computes proportion of nearest neighbors preservation in reduced dimension in comparison to high dimension}
\usage{
computeDimRed_KNNQC(CYTdata, knn = 100, subsetSize = 10000, markers = NULL)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{knn}{a numeric being the number of nearest neighbors to compute. Defaults to 100.}

\item{subsetSize}{a numeric being the sample size used for computation. Defaults to 10000, in order to avoid risks of memory shortage,}

\item{markers}{a character vector of the markers to consider for the knn. Defaults to the markers used to compute the dimensionnality reduction (recommended).}
}
\value{
a numeric proportion
}
\description{
Computes the fraction of k-nearest neighbors in the original high-dimensional data that are preserved as k-nearest neighbors in the embedding
The average is computed across all n points. KNN quantifies preservation of the local, or microscopic structure.
Please note that results will vary greatly depending on parameters knn and subsetSize. Please refer to Wang et al., 2023 (https://www.nature.com/articles/s41467-023-37478-w) for quality control of dimension reduction method in cytometry.
}
