% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Metaclustering.R
\name{computeMSI}
\alias{computeMSI}
\title{Computes median or mean for a set of marker expression}
\usage{
computeMSI(
  CYTdata,
  markers = NULL,
  population = NULL,
  level = c("clusters", "metaclusters"),
  typeMSI = c("median", "mean"),
  rescaleBounds = c(0.05, 0.95),
  scale = TRUE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a vector object listing marker names}

\item{population}{a character value containing the identifiers of the population(s) (clusters or metaclusters) to study}

\item{level}{a character value indicating whether to compute MSI on the clusters or on the metaclusters. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{typeMSI}{a character value to choose between "median" (default) and "mean".}

\item{rescaleBounds}{a numeric value providing the value of the first quantile and last quantile for the rescaled values, to be passed to rescale.exprs()}

\item{scale}{Logical. If TRUE, rescales the value of each marker's expression to the bounds indicated in rescaleBounds}
}
\value{
a dataframe
}
\description{
Computes median or mean expression for a set of given markers, per cluster or per metacluster
}
