% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering_QC.R
\name{computePopulation_SizeQC}
\alias{computePopulation_SizeQC}
\title{Computes the percentage of clusters with sufficient number of cells}
\usage{
computePopulation_SizeQC(
  CYTdata,
  population = NULL,
  level = c("clusters", "metaclusters"),
  thSize = 50,
  generateHeatmap = FALSE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'
level = c("clusters", "metaclusters")}

\item{population}{a character vector containing the identifiers of the clusters/metaclusters to study. Defaults to all of them.}

\item{level}{a character value indicating whether to display clusters or metaclusters. Possible values are: "clusters" (default), "metaclusters".}

\item{thSize}{a numeric value providing the minimum number of cells needed for a cluster to be considered a small cluster (default value = 50)}

\item{generateHeatmap}{logical. If TRUE, also generates a heatmap with the clusters passing the threshold for each sample and in total. Defaults to FALSE}
}
\value{
a list containing QC size features:
- cellcount: a cellcount dataframe
- vectorScores: a boolean vector indicating for each cluster if it passes the threshold. 
- thSize: the threshold used
- score: the percentage of clusters that pass the threshold
- heatmap: a ggplot of the QC heatmap.
}
\description{
This function aims to compute and show cell clusters/metaclusters having a number of associated cells greater than a specific threshold.
Cluster cell count is computed for all samples at once. The optional heatmap, however, will show every cluster/sample pair.
}
