% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering_QC.R
\name{computeQCuniform}
\alias{computeQCuniform}
\title{Computes the percentage of clusters with uniform phenotypes}
\usage{
computeQCuniform(
  CYTdata,
  markers = NULL,
  population = NULL,
  level = c("clusters", "metaclusters"),
  thPvalue = 0.05,
  thIQR = 2,
  thGood = 0.75,
  NbSim = 2000,
  generateHeatmap = TRUE,
  sortQCbyMarkers = TRUE,
  generateEcdf = FALSE,
  parallelize = FALSE,
  nbCores = NULL
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a character vector indicating the markers to use in the QC. 
By default, all of them are used, but it is recommended to use only the markers used for  clustering.}

\item{population}{a character vector containing the identifiers of the clusters/metaclusters to compute the QC on. Defaults to all of them.}

\item{level}{a character value indicating whether to compute the QC on clusters or metaclusters. Possible values are: "clusters" (default), "metaclusters".}

\item{thPvalue}{a numeric value providing the p-value threshold for the test (unimodal if pvalue > thPvalue). Defaults to 0.05}

\item{thIQR}{a numeric value providing the IQR (interquartile range) threshold to assume a distribution as narrow enough. Default to 2}

\item{thGood}{a numeric value providing the proportion of markers that need to pass the QC in order to consider that a cluster passes the QC}

\item{NbSim}{an integer. Number of replicates for the Monte Carlo simulation. Defaults to 2000.}

\item{generateHeatmap}{logical. If TRUE (default), a heatmap is generated with the QC.}

\item{sortQCbyMarkers}{logical. If TRUE (default), markers will be sorted on the heatmap by the number of clusters that pass the QC for that marker.}

\item{generateEcdf}{logical. If TRUE, an empirical cumulative distribution is generated with the QC, useful for comparing thGood values. CURRENTLY NON FUNCTIONNAL.}

\item{parallelize}{logical. If TRUE, calculation of the test is done in parallel on several cores. If you get an "impossible to find function "%dopar%" error, try library(foreach)}

\item{nbCores}{an integer. Number of processor cores to use. If NULL, uses one less cores than available to leave computational power for using the computer.}
}
\value{
a list containing QC uniform features:
- Pvalue: a dataframe containing the pvalues for the dip test
- IQR: a dataframe containing interquartile range
- Uniform: a dataframe indicating which cluster/marker pair passes the QC for both uniformity and IQR
- goodScores: a named vector giving the percentage of markers passing the QC for each cluster
- Score: the overall percentage of clusters passing the QC
- thPvalue: the thPvalue parameter
- thIQR the thIQR parameter
- thGood: the thGood parameter
- heatmap: a ggplot of the QC heatmap. 
- ecdf: a ggplot of the empirical cumulative distribution
}
\description{
This function aims to identify and show cell clusters having a uniform phenotype, defined as an unimodal expression and a narrow range of expression for a selected set of markers

 @details

- Check unimodal expression: testing unimodal distribution of markers with Hartigan's dip test.
- Check the range of expression: check if the interquartile range of the specified markers is below a specified threshold.
}
