% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_FCS_Markers_CYTdata.R
\name{createCYTdata}
\alias{createCYTdata}
\title{Imports of cell expression profiles from FCS files}
\usage{
createCYTdata(
  files,
  channels,
  transformList = NULL,
  rawData = FALSE,
  Ndownsampling = NULL,
  ...
)
}
\arguments{
\item{files}{a character vector specifying the path of the tab-separated or FCS files to load}

\item{channels}{a character vector providing the channels to import from FCS/text files. By default, only de common channels to all the FCS files are kept}

\item{transformList}{an S4 object of class transformList with features integrated (transform function, channels to transform, parameters). See Flowcore package : https://www.bioconductor.org/packages/release/bioc/manuals/flowCore/man/flowCore.pdf}

\item{rawData}{a boolean specifying if raW data should be also imported and stored into CYTdata's slot "rawExpression".}

\item{Ndownsampling}{an integer specifying the number of cells per sample for downsampling}

\item{...}{argument to pass onto method read.FCS from Flowcore package}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to import acquired cell events into a CYTdata object.

Input files must be FCS files.
Different transformations can be applied such as logicle, arcsinh. If yes, parameters are added "..."
Cell marker having technical or biological biaises can be excluded during the import.
}
