% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{CYTdata-class}
\alias{CYTdata-class}
\alias{CYTdata}
\title{CYTdata class definition}
\description{
The CYTdata object is a S4 object containing all cytometry expressions.
}
\section{Slots}{

\describe{
\item{\code{samples}}{a character vector containing the names of the biological samples}

\item{\code{matrix.expression}}{a data.frame containing the marker expressions of each cells (data to analyse)}

\item{\code{metadata}}{a data.frame containing the metadata associated to each sample}

\item{\code{raw.matrix.expression}}{a data.frame containing the marker expressions of cells from the whole dataset before transformation process (raw data)}

\item{\code{Clustering}}{a S4 object of class 'Clustering' containing the clustering step-related informations}

\item{\code{Metaclustering}}{a S4 object of class 'Metaclustering' containing the metaclustering step-related informations}

\item{\code{DimReduction}}{a S4 object of class 'DimReduction' containing the dimensionality reduction step-related informations}

\item{\code{DiffAbundanceAnalysis}}{a data.frame containing the statistics of comparison computed with 'run.DiffAbundanceAnalysis' function}

\item{\code{Kinetic}}{a S4 object of class 'Kinetic'}
}}

