% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_Gating_Merging.R
\name{CYTdataGating}
\alias{CYTdataGating}
\title{Gate cells in a CYTdata object}
\usage{
CYTdataGating(
  CYTdata,
  subset,
  level = c("samples", "clusters", "metaclusters")
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{subset}{a character vector containing the identifier of the clusters, metaclusters or samples to keep}

\item{levels}{a character value indicating whether to gate according to "samples" (default), "clusters" or "metaclusters".}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Gate cells in a CYTdata object according to clusters, metaclusters or samples and create new CYTdata object.
Dimension reduction coordinates are not kept
Gating can be performed according to a specific metadata by setting subset to the name of the samples corresponding to the desired metadata, in the form subset = rownames(filter(CYTdata@metadata, Cohort == "GC" & Timepoint %in% c("BSL", "D1", "D3", "D6")))
}
