% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_Gating_Merging.R
\name{CYTdataMerging}
\alias{CYTdataMerging}
\title{Merge two CYTdata objects}
\usage{
CYTdataMerging(
  CYTdata1,
  CYTdata1Name = "CYTdata1",
  CYTdata2,
  CYTdata2Name = "CYTdata2",
  markers = NULL,
  checkDuplicates = FALSE,
  mergePopulation = TRUE,
  force = FALSE,
  sameSplMetadata = TRUE,
  addMetadata = TRUE,
  renameforAll = TRUE
)
}
\arguments{
\item{CYTdata1}{a S4 object of class 'CYTdata'}

\item{CYTdata1Name}{a character vector containing the name for the first CYTdata object. Defaults to "CYTdata1"}

\item{CYTdata2}{a S4 object of class 'CYTdata'}

\item{CYTdata2Name}{a character vector containing the name for the second CYTdata object. Defaults to "CYTdata2"}

\item{markers}{a character vector containing the name of the markers to keep in the cell expression matrix. Defaults to all common markers.}

\item{checkDuplicates}{logical. If TRUE, checks for cell duplicates (i.e., cells with the exact same marker expression). May be slow. Defaults to FALSE.}

\item{mergePopulation}{logical. If TRUE (default), clusters and metaclusters with identical names in both CYTdata objects will be merged. If FALSE, a suffix will be appended to all or identical (see renameforAll) clusters/metaclusters depending on their original CYTdata.}

\item{force}{logical. If TRUE, CYTdata objects will be merged even if one of them contains no metadata slot. If FALSE (default) and one of their metadata slots is empty, will return an error.}

\item{sameSplMetadata}{logical. If TRUE (default), both CYTdatas are expected to have the same sample names. If FALSE, they must have no samples in common.}

\item{addMetadata}{logical. If TRUE (default), adds a new metadata column indicating the original CYTdata's name. Only works if sameSplMetadata = FALSE.}

\item{renameforAll}{logical. If TRUE (default) and mergePopulation = FALSE, all clusters/metaclusters will be appended the name of their original CYTdata, even if they have no cluster/metacluster with the same name in the other CYTdata.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Merge two CYTdata objects.
}
