% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Metadata_Samples.R
\name{CYTfilterBool}
\alias{CYTfilterBool}
\title{Filter cells by metadata and return a boolean}
\usage{
CYTfilterBool(CYTdata, metadata, condition)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{metadata}{a character value of length 1 indicating the metadata of interest (e.g., "Timepoint" or "Individual)}

\item{condition}{a character vector indicating the condition(s) of interest (e.g., c("Baseline", "Day1", "Day7")). Can be of any length > 0.}
}
\value{
a boolean vector
}
\description{
This function returns a boolean vector indicating whether or not each cell in the CYTdata matches a specific metadata condition.
}
