% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Metadata_Samples.R
\name{CYTfilterIdx}
\alias{CYTfilterIdx}
\title{Filter cells by metadata and return indexes}
\usage{
CYTfilterIdx(CYTdata, metadata, condition)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{metadata}{a character value of length 1 indicating the metadata of interest (e.g., "Timepoint" or "Individual)}

\item{condition}{a character vector indicating the condition(s) of interest (e.g., c("Baseline", "Day1", "Day7")). Can be of any length > 0.}
}
\value{
a numerical vector
}
\description{
This function returns indexes of cells in a CYTdata object matching a specific metadata condition.
}
