% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytopipeRdev.R
\docType{package}
\name{CytopipeRdev}
\alias{CytopipeRdev-package}
\alias{CytopipeRdev}
\title{CytopipeRdev}
\description{
This package is intended primarily for use at the IDMIT lab (https://jacob.cea.fr/drf/ifrancoisjacob/english/Pages/Departments/IDMIT.aspx). It has been developped initially by Paul MAZET, then continued by Cyril ETIENNE and Aurélien MARC.
}
\details{
It offers simplified functions for the analysis of cytometry datasets, notably clustering of cells, visualization, and differential analysis

Any questions regarding the use of the package can be adressed to cyril.etienne@cea.fr or the IBI lab of IDMIT.
}
