% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{DimReduction-class}
\alias{DimReduction-class}
\alias{DimReduction}
\title{DimReduction class definition}
\description{
The DimReduction object is a S4 object containing all the data concerning the DimReduction step performed on a 'CYTdata' object
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{a data.frame containing the coordinates data points in the reduced data space}

\item{\code{optional_parameters}}{a list containing the parameters used for DimReduction computation (especially the arguments of 'run.DimReduction' function : seed, markers, type, etc.)}
}}

