% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_Sampling.R
\name{downsampling.density}
\alias{downsampling.density}
\title{Density-based downsampling (Internal)}
\usage{
downsampling.density(
  indexes.to.downsample,
  target.size,
  whole.data.exprs,
  exclude.pctile,
  compensate.uniform,
  ncores
)
}
\arguments{
\item{indexes.to.downsample}{a numeric vector specifying the indexes of a sample's events in the dataset}

\item{target.size}{a numeric value specifying the number of events targeted as downsampled, is smaller than the number of indexes to downsample}

\item{whole.data.exprs}{a data.frame containing the entire dataset : all the markers' expression (used for density computation)}

\item{exclude.pctile}{a numeric value, between 0 and 1, specifying the quantile value used to determine the exclusion threshold for density values}

\item{compensate.uniform}{a boolean value : Density based downsampling doesn't permit us to get the exact size specified by "target.size". So, do we add a step to adjust randomly the donsampled indexes ?}

\item{ncores}{numeric, number of processor cores to use.}
}
\value{
downsampled.indexes a numeric vector specifying the indexes downsampled
}
\description{
Downsample indexes representing sample's events with a density-based method (spade)
}
