% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_FCS_Markers_CYTdata.R
\name{export.CYTdata.expression}
\alias{export.CYTdata.expression}
\title{Exports cell expression matrix to TSV or FCS files}
\usage{
export.CYTdata.expression(
  CYTdata,
  outputDir = ".",
  markers = NULL,
  samples = NULL,
  population = NULL,
  level = c("clusters", "metaclusters"),
  concatenate.samples = FALSE,
  extension = c("FCS", "TXT")
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{outputDir}{a character value providing the path of folder where the output files are stored}

\item{markers}{a character vector containing the markers of which the expression is to export.}

\item{samples}{a character vector containing the names of biological samples to export. By default, all samples are extracted.}

\item{population}{a character vector containing the identifiers of the population to export.
If set to NULL (default), the selection by population is not used (useful if no clustering/metaclustering step has been performed)}

\item{level}{a character value indicating the type of population extracted. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{concatenate.samples}{If set to TRUE, sample expression are concatenated into the same FCS files. By default, one FCS is written file by sample.}

\item{extension}{a character being the type of file into which data is exported. Possible values are: "FCS", "TXT" (default =  "FCS")}
}
\value{
none
}
\description{
Exports cell expression matrix from a CYTdata object to a tab-separated or FCS files. One FCS file by sample (according to sample CYTdata sample vector)
Cell expression profiles can be exported for a set of given samples and for a set of given cell population
}
