% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_DimReduction.R
\name{export.DimReduction}
\alias{export.DimReduction}
\title{Export a S4 object of class 'DimReduction' from a CYTdata object}
\usage{
export.DimReduction(CYTdata)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}
}
\value{
a S4 object of class 'DimReduction'
}
\description{
Strictly equivalent to "CYTdata@DimReduction"
}
