% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Clustering.R
\name{exportClustering}
\alias{exportClustering}
\title{Exports an  object of class 'Clustering' from a CYTdata object}
\usage{
exportClustering(CYTdata)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}
}
\value{
a S4 object of class 'Clustering'
}
\description{
Exports an  object of class 'Clustering' from a CYTdata object. Equivalent to "CYTdata@Clustering"
}
