% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_Expression_Viewer.R
\name{expression.boxplot}
\alias{expression.boxplot}
\title{Plot a boxplot of marker expression}
\usage{
expression.boxplot(CYTdata, markers = NULL, rawData = FALSE)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a character vector containing the names of biological markers to use. By default, all markers are plotted}

\item{rawData}{a boolean value specifying if data come from raw.matrix.expression slot of CYTdata object or if data come from matrix.expression slot.
By default, data from matrix.expression is used}
}
\value{
a ggplot2 object
}
\description{
This function plots a CYTdata object's expression data.
Warning : VERY slow.
}
