% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_FCS_Markers_CYTdata.R
\name{getChannelsMarkersFCS}
\alias{getChannelsMarkersFCS}
\title{Get the markers (and channels associated) common to a set of FCS or TXT files}
\usage{
getChannelsMarkersFCS(files, removeTime = TRUE, verbose = FALSE)
}
\arguments{
\item{files}{a character vector specifying the path of the tab-separated or FCS files to load}

\item{removeTime}{a boolean specifying whether to remove the Time column from the data}

\item{verbose}{a boolean specifying whether to return additionnal information}
}
\value{
a named character vector
}
\description{
This function aims at importing marker and channel names from a set of FCS files

Input files must be FCS or TXT files.
Only common channels and markers are returned
}
