% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Population_Palette.R
\name{getMetaclustersMembership}
\alias{getMetaclustersMembership}
\title{Get the names of the clusters belonging to each metacluster.}
\usage{
getMetaclustersMembership(CYTdata, format = c("list", "data.frame"))
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{format}{a string specifying the returned data's format. Possible values are "list" (default) and "data.frame"}
}
\value{
a list or dataframe indicating the clusters associated with each metacluster
}
\description{
Get the names of the clusters belonging to each metacluster.
}
