% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_Abundance_Viewer.R
\name{getRelativeAbundance}
\alias{getRelativeAbundance}
\title{Get the abundance of a specific cell population}
\usage{
getRelativeAbundance(
  CYTdata,
  population = NULL,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  Yvalue = c("percentage", "absolute")
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{population}{a character vector containing the identifiers of the population for which to return abundance or cell count. By default, all of them are used, which will return 100% if Yvalue is set to "percentage".}

\item{level}{a character value indicating the type of population to study. Possible values are: "clusters" (default) and "metaclusters".}

\item{samples}{a character vector containing the names of biological samples to consider. By default, all samples are used}

\item{Yvalue}{a character value containing the parameters to use. Possible value are "percentage" (default) or "absolute" (i.e., cell count).}
}
\value{
a data frame with abundance per sample, along with the metadata of each sample
}
\description{
This function returns the relative or absolute abundance of specific cell populations (i.e., clusters or metaclusters)
}
