% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_Abundance_Viewer.R
\name{getRelativeAbundanceThreshold}
\alias{getRelativeAbundanceThreshold}
\title{Gets number of cells above an expression threshold for a given marker.}
\usage{
getRelativeAbundanceThreshold(
  CYTdata,
  marker,
  th,
  population = NULL,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  Yvalue = c("percentage", "absolute")
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{marker}{a character vector containing the name of the marker to use. Must be included in CYTdata's expression matrix's colnames.}

\item{th}{a numeric value giving the threshold to compare the marker expression to. If the data are transformed (e.g., an arcsinh), then the threshold must be transformed as well.}

\item{population}{a character vector containing the identifiers of the clusters/metaclusters to study.}

\item{level}{a character value containing the levels to display in the plot, either "clusters" (default) or "metaclusters"}

\item{samples}{a character vector containing the names of the biological samples to use. By default, all samples are used.}

\item{Yvalue}{a character value containing the variable to be studied. Possible values are percentage (default) or absolute.}
}
\value{
a dataframe containing the number/percentage of cells per sample for the given pop, along with collated metadata
}
\description{
This function aims to compute cellcount or abundance across samples for a given population (one or several cluster or metacluster), counting only cells for which the expression of a given marker is above a given threshold (e.g., a positivity thressold for IdU+ cells. )
As of now, percentage results are computed over the total number of cells, not the total number of marker+ cells.
}
