% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Metadata_Samples.R
\name{getSamplesMetadata}
\alias{getSamplesMetadata}
\title{Get the sample names associated to a specific metadata}
\usage{
getSamplesMetadata(CYTdata, metadataCondition)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{metadataCondition}{a string specifying the metadata condition chosen. The string has to be the name of the column (in metadata data.frame) containing the condition chosen}
}
\value{
a list containing the samples for each value of condition chosen
}
\description{
This function takes a given metadata (e.g. "Timepoint" or "Individual") and returns a list giving the sample names associated with each level of the condition, e.g, each sample associated to "Timepoint" = "BSL", "D1", etc.
}
