% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Population_Palette.R
\name{heritatePalette}
\alias{heritatePalette}
\title{Give clusters the same color as metaclusters}
\usage{
heritatePalette(CYTdata, population = NULL, addVariation = FALSE)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{population}{a character vector containing the identifiers of the clusters for which to change the color. If NULL, (default) all clusters are used}

\item{addVariation}{logical. If TRUE, some brightness adjustments will be made to the metacluster colors to be able to distinguish the clusters while still retaining the hue of the metacluster. Defaults to FALSE.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function  changes the color of the clusters to match that of their respective metaclusters.
}
