% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Clustering.R
\name{importClustering}
\alias{importClustering}
\title{Import object of class 'Clustering' to put in CYTdata object}
\usage{
importClustering(CYTdata, importedClusteringObject, checkOverwrite = TRUE)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{importedClusteringObject}{a S4 object of class 'Clustering' to put in CYTdata object}

\item{checkOverwrite}{a boolean value indicating whether to check if a Metaclustering has already been performed on the CYTdata object. 
If TRUE, it will check and ask for confirmation, if FALSE, any previous clustering will be overwritten by default.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Import object of class 'Clustering' to put in CYTdata object
}
