% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_DimReduction.R
\name{importDimReduction}
\alias{importDimReduction}
\title{Import an object of class 'DimReduction' to put into a CYTdata object}
\usage{
importDimReduction(CYTdata, importedDimReductionObject, checkOverwrite = TRUE)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{importedDimReductionObject}{a S4 object of class 'DimReduction' to put in CYTdata object}

\item{checkOverwrite}{a boolean, if TRUE (default), will pause and display a warning if a DimReduction is already present in the CYTdata and must be overwritten.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Import an object of class 'DimReduction' to put into a CYTdata object
}
