% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Metaclustering.R
\name{importMetaclustering}
\alias{importMetaclustering}
\title{Import object of class 'Metaclustering' to put in CYTdata object}
\usage{
importMetaclustering(
  CYTdata,
  importedMetaclusteringObject,
  checkOverwrite = TRUE,
  val = TRUE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{importedMetaclusteringObject}{a S4 object of class 'Metaclustering' to put in CYTdata object}

\item{checkOverwrite}{a boolean value indicating whether to check if a Metaclustering has already been performed on the CYTdata object. 
If TRUE, it will check and ask for confirmation, if FALSE, any previous clustering will be overwritten by default.}

\item{val}{a boolean value. If TRUE, performs validation of the CYTdata object (makeValid function) after importing the Metaclustering.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Import object of class 'Metaclustering' to put in CYTdata object
}
