% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Metadata_Samples.R
\name{importMetadata}
\alias{importMetadata}
\title{Assigns meta-information about biological samples to a CYTdata object}
\usage{
importMetadata(CYTdata, metadataDf, checkOverwrite = TRUE)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{metadata}{a dataframe containing meta-information about the biological samples.
The columns must contain, at least, a column named "Timepoint" and an other named "Individual".
The rownames have to be the biological samples, thus the number rows has to be equal to the number of samples.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to attach meta-information to each biological sample present in a CYTdata object.

Especially, the following meta-information of each sample can be specified for subsequent analyses.
- The biological individual
- The biological condition (groups, vaccines, studies, etc.)
- The timepoint
Timepoint and Individual data must be specified.
}
