% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{Kinetic-class}
\alias{Kinetic-class}
\alias{Kinetic}
\title{Kinetic class definition}
\description{
The Kinetic object is a S4 object containing all the data concerning a kinetic clustering step performed on a 'CYTdata' object
}
\section{Slots}{

\describe{
\item{\code{families}}{a data.frame containing the identifiers of metaclusters kinetic families computed with 'run.Kinetic' function}

\item{\code{optional_parameters}}{a list containing the parameters used for kinetic clustering (especially the arguments of 'run.Kinetic' function : N.kinetics, sample.subset, etc.)}
}}

