% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\name{MakeValid}
\alias{MakeValid}
\title{Checks validity of all elements of a CYTdata object}
\usage{
MakeValid(CYTdata, verbose = TRUE)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{verbose}{whether to display additional messages}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Checks validity of all elements of a CYTdata object (préciser les étapes ?)
}
