% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Metaclustering.R
\name{manualMetaclustering}
\alias{manualMetaclustering}
\title{Perform manual metaclustering of the clusters}
\usage{
manualMetaclustering(
  CYTdata,
  belongMetaclustering,
  completeWith = c("clusters", "metaclusters"),
  checkOverwrite = TRUE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{belongMetaclustering}{A list of manually set metaclusters with their associated cluster names.
For example : belongMetaclustering = list("Monocytes" = c("cluster1","cluster3","cluster4","cluster5"), Lymphocytes = c("cluster2","cluster6","cluster7"), etc)}

\item{completeWith}{a character value indicating how unassigned clusters should be treated. 
If "clusters", unassigned clusters will each be assigned to a metacluster named after them. 
If "metaclusters",unassigned clusters will keep their previous metacluster assignment (requires metaclustering to have been completed before).}

\item{checkOverwrite}{a boolean value indicating whether to check if a Metaclustering has already been performed on the CYTdata object. If TRUE, it will check and ask for confirmation, if FALSE, any previous clustering will be overwritten by default.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function takes of list of metacluster with their associated clusters to integrate metaclustering to a CYTdata object. Useful if all clusters have been manually assigned to a specific cell population that can be stored in the Metaclustering slot of the CYTdata object.
}
