% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_Expression_Viewer.R
\name{markerDensity}
\alias{markerDensity}
\title{Plot distribution of marker expression}
\usage{
markerDensity(CYTdata, markers, samples = NULL, group.metadata = NULL)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a character vector containing the names of a biological marker to use. Currently, only one of them can be used.}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{group.metadata}{a character value containing the names of the metadata to use to split the plot. Possible values are colnames of CYTdata's metadata slot. by default, it is set to NULL, which means that markers densities are not grouped by metadata condition.}
}
\value{
void
}
\description{
This function aims to plot the distribution of marker expression for selected samples. Marker expression can be further divided according to any kind of metadata.
}
