% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_Expression_Viewer.R
\name{markersScatterplot}
\alias{markersScatterplot}
\title{Plot scatterplot of 2 marker's expression}
\usage{
markersScatterplot(
  CYTdata,
  coupleMarkers,
  samples = NULL,
  level = c("clusters", "metaclusters"),
  population = NULL,
  rawData = FALSE,
  colorBy = c("none", "clusters", "metaclusters", "density", "metadata"),
  colorMetadata = "Timepoint",
  neighboursDensity = 100,
  colorNone = NULL,
  plotLinearRegression = FALSE,
  downsamplingNumber = NULL,
  ylim = NULL,
  xlim = NULL
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{coupleMarkers}{a character vector providing the marker represented on each axis (axis x then axis y)}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{level}{a character vector indicating whether to select "clusters" (default) or "metaclusters" for the population argument.}

\item{population}{a character vector indicating the identifiers of the clusters/metaclusters to compare. Defaults to all of them.}

\item{rawData}{a boolean value specifying if data comes from raw.matrix.expression slot of CYTdata object or if data comes from matrix.expression slot. By default, data from matrix.expression is used}

\item{colorBy}{a character value providing the kind of coloration to apply to the plot. Possible values are : none (default), "clusters", "metaclusters", "density", "metadata". "density" will display a density plot.}

\item{colorMetadata}{a character value providing the name of the metadata according to which the points should be colors if colorBy = "metadata".}

\item{neighboursDensity}{a numeric value providing the size of the grid for density calculations (effectively, the breaks of a 2D histogram). Defaults to 100.}

\item{colorNone}{a character value providing the name or hexadecimal code for point color if colorBy = "none".}

\item{plotLinearRegression}{logical. If TRUE, a linear regression is performed and displayed on the plot. Please note no test is performed. Defaults to FALSE.}

\item{downsamplingNumber}{a numeric value providing a downsampling target for displaying the points. Defaults to NULL, which is not recommended for large CYTdata objects.}

\item{ylim}{a numeric of length 2 providing lower and upper bounds for the y axis}

\item{xlim}{a numeric of length 2 providing lower and upper bounds for the x axis}
}
\value{
a ggplot2 object
}
\description{
This function permits to plot a CYTdata object's expression data over scatterplot with 2 markers choosen
}
