% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_CYTdata_classes.R
\docType{class}
\name{Metaclustering-class}
\alias{Metaclustering-class}
\alias{Metaclustering}
\title{Metaclustering class definition}
\description{
The Metaclustering object is a S4 object containing all the data concerning the Metaclustering step performed on a 'CYTdata' object
}
\section{Slots}{

\describe{
\item{\code{metaclusters}}{a vector containing the identifiers of cell metaclusters (phenotypic families) as a factor}

\item{\code{cellcount}}{a data.frame containing the number of cells associated to each metacluster for each sample}

\item{\code{abundance}}{a data.frame containing the percentage of cells associated to each metacluster for each sample}

\item{\code{palette}}{TO DO}

\item{\code{optional_HierarchicalHeatmap}}{TO CHECK a list containing all the features about heatmap representing phenotypes of clusters and families}

\item{\code{optional_parameters}}{TO CHECK a vector containing the parameters used for the identification of the cell clusters (especially the argument of 'run.Clustering' function : seed, markers, architecture, type, etc.)}
}}

