% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_Sampling.R
\name{performDownsampling}
\alias{performDownsampling}
\title{Performs the downsampling of events}
\usage{
performDownsampling(
  CYTdata,
  type = c("uniform", "density"),
  sizeType = c("bySplPercent", "bySplSmall", "bySplAbsolute", "byOverallAbsolute"),
  sizeAbsolute = 1000,
  sizePercent = 1/5,
  seed = 42,
  density.exclusion.pctile = 0.01,
  density.compensate = FALSE,
  density.ncores = 1
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{type}{a string value specifying the type of downsampling used, "uniform" (default) or "density" (density-based). Density based downsampling may be slow.}

\item{sizeType}{a string value specifying the way the number of downsampled events is determined ("bySplPercent" (default) to use a fraction of each sample's original size, "bySplSmall" to use the smallest sample's size, "bySplAbsolute" to use an absolute size per sample, "byOverallAbsolute" to use an absolute size for the whole dataset)}

\item{sizeAbsolute}{a numeric value specifying the target number of cells, either by sample (if sizeType = "bySplAbsolute") or overall (if sizeType = "byOverallAbsolute"). Defaults to 1000.}

\item{sizePercent}{a numeric value specifying the target fraction of sample size for downsampling (expected if sizeType = "bySplPercent"). Defaults to 0.2.}

\item{seed}{a numeric value: the seed for RNG.}

\item{density.exclusion.pctile}{a numeric value, between 0 and 1, specifying the quantile value used to determine the exclusion threshold for density values. Defaults to 0.01}

\item{density.compensate}{logical. Density based downsampling doesn't allow getting the exact size specified by "target.size". If TRUE, a step is added to adjust randomly the downsampled indexes. Defaults to FALSE.}

\item{density.ncores}{a numeric value specifying the number of logical processor used for parallel computation in the case of density based downsampling}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Perform the downsampling, using uniformly-based or density-based random selections, of events
}
