% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_Sampling.R
\name{performUpsampling}
\alias{performUpsampling}
\title{Performs the upsampling of downsampled events}
\usage{
performUpsampling(
  CYTdata,
  newCYTdata,
  markers = NULL,
  type = c("KNN", "RF", "Logistic.regression")
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{newCYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a character vector providing the markers to use to perform upsampling}

\item{type}{a character value containing the type of the method to apply. Possible values are: "KNN", "RF", "Logistic.regression" (default = KNN)}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to perform the upsampling of downsampled events events based on an existing CYTdata object and existing cell events stored in tab-separated or FCS files.

Importantly, the identification of cell clusters must have been performed prior to this operation.
}
