% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_Abundance_Viewer.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Plots cell cluster abundances using a boxplot representation}
\usage{
plotBoxplot(
  CYTdata,
  population,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  observationMetadata = "Timepoint",
  groupMetadata = NULL,
  pointMetadata = NULL,
  Yvalue = c("percentage", "absolute"),
  computePval = FALSE,
  test.statistics = c("wilcox.test", "t.test"),
  paired = FALSE,
  corrMethod = "holm"
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{population}{a character vector containing the names of the population to use.}

\item{level}{a character value containing the levels to display in the plot, either "clusters" (default) or "metaclusters"}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{observationMetadata}{a character value containing the biological condition to use for the boxplot groups. 
The name must match a column of the CYTdata's metadata slot. By default, "Timepoint" is used.}

\item{groupMetadata}{a character value containing the biological condition (as for observationMetadata) to use for a secondary division of the boxplot groups. Defaults to NULL.}

\item{pointMetadata}{a character value containing the biological condition (as for observationMetadata) to use for displaying point shape. Defaults to NULL.}

\item{Yvalue}{a character value containing the quantitative variable to display. Possible values are "percentage" (the relative abundance, default) or "absolute." (the cell count)}

\item{computePval}{a boolean value indicating if statistics should be computed.}

\item{test.statistics}{a character value providing the type of statistical test to use. Possible values are: 'wilcox.test' or 't.test'}

\item{paired}{a boolean value indicating if a paired or unpaired comparison should be applied}

\item{corrMethod}{a character value providing the type of correction for the p.value. Please refer to rstatix::wilcox_test or rstatix::t_test for details.}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize and compare the cell cluster abundances for each biological condition using boxplot and jitter representations.

The abundance of a specific cell cluster or a set of cell clusters can be displayed.
The representation can be restricted to a specific set of samples.
Moreover, boxplot can be constructed based on sample meta information.
Statistics can be computed for all comparisons.
}
