% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Clustering_Viewer.R
\name{plotCorrDiagram}
\alias{plotCorrDiagram}
\title{Plots a diagram showing correlated markers}
\usage{
plotCorrDiagram(
  CYTdata,
  markers = NULL,
  population = NULL,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  method = c("pearson", "spearman"),
  palette = NULL
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{markers}{a character vector containing the name of the markers to analyse}

\item{population}{a character vector containing the identifiers of the population to use. By default, all the population are used}

\item{level}{a character value indicating the type of population plotted. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{samples}{a character vector containing the name of the samples to consider}

\item{method}{a character value indicating the name of correlation method to use, "pearson" (default) or "spearman"}

\item{palette}{optionnal; a vector of color names to use for the gradient.}
}
\value{
a list with:
 - corPlot: a ggplot2 object, the correlation diagram
 - corMatrix: the correlation matrix
}
\description{
This function aims to visualize the pairwise co-expression between all markers.
Each tile corresponds to the co-expression between two markers and is gradient-colored based on the Pearson or Spearman correlation
}
