% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_DimReduction_Viewer.R
\name{plotDimRedGradient}
\alias{plotDimRedGradient}
\title{Visualize 2D reduced cell data according to marker expression}
\usage{
plotDimRedGradient(
  CYTdata,
  marker,
  population = NULL,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  bounds = c(0.05, 0.95),
  relativeGradient = FALSE,
  paletteGradient = NULL
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{marker}{a character vector containing the name of the biological marker to plot}

\item{population}{a character vector containing the identifiers of the population to plot. By default, all the population are used. Non included populations will still be ploted, but in grey.}

\item{level}{a character value indicating the type of population filtered by the population argument. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{samples}{a character vector containing the names of the biological samples to plot. Non included samples will still be ploted, but in grey.}

\item{bounds}{a numeric providing the two quantiles that will be used to bind the expression of the marker. Defaults to c(0.05, 0.95).}

\item{relativeGradient}{a boolean. If TRUE, marker expression will be bound according to the marker's expression in the selected populations and/or samples. If FALSE (default), the entire cell population will be used.}

\item{paletteGradient}{a character vector containing the colors to use for the gradient scale. Defaults to a blue/red gradient.}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize cells in the embedded 2D space, by coloring them with the expression of a given marker.
This representation can be used on a CYTdata object for which a dimensionality reduction operation has been computed.
}
