% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_DimReduction_Viewer.R
\name{plotDimRedPopulations}
\alias{plotDimRedPopulations}
\title{Visualize 2D reduced cell data according to clustering or metaclustering}
\usage{
plotDimRedPopulations(
  CYTdata,
  population = NULL,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  printCentroid = TRUE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{population}{a character vector containing the identifiers of the population to plot. By default, all the population are used. Non included populations will still be ploted, but in grey.}

\item{level}{a character value indicating the type of population filtered by the population argument. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{samples}{a character vector containing the names of the biological samples to plot. Non included samples will still be ploted, but in grey.}

\item{printCentroid}{a boolean. If TRUE (default), displays the name of each population at that population's centroid position.}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize cells in the embedded 2D space, by coloring them with the expression of a given marker.
This representation can be used on a CYTdata object for which a dimensionality reduction operation has been computed.
}
