% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_DimReduction_Viewer.R
\name{plotDimReduction}
\alias{plotDimReduction}
\title{Visualize 2D reduced cell data according to density and metadata}
\usage{
plotDimReduction(
  CYTdata,
  samples = NULL,
  population = NULL,
  level = c("clusters", "metaclusters"),
  plotType = c("density", "metadata"),
  title = "Dimensionnality reduction view",
  metadata = "Timepoint"
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{samples}{a character vector containing the names of the biological samples to plot. Non included samples will still be ploted, but in grey.}

\item{population}{a character vector containing the identifiers of the population to plot. By default, all the population are used. Non included populations will still be ploted, but in grey.
Please note that this filters the cells that are ploted but does not affect their colors, for that, please see plotDimRedPopulations}

\item{level}{a character value indicating the type of population filtered by the population argument. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{plotType}{a character value indicating the type of coloring to apply. 
Possible values are "density" (default, the points are colors according to cell density as if in a 2D histogram) and "metadata" (the points are coloured according a specified metadata)}

\item{title}{a character value providing the title for the graph. Defaults to "Dimensionnality reduction view"}

\item{metadata}{a character vector containing the type of metadata to study, if plotType = "metadata".}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize cells in the two dimensional embedded space after dimensionality reduction has been performed. 
This representation can be used on a CYTdata object for which a dimension reduction has been computed
Cells can be colored according to any kind of metadata stored in the CYTdata object, or according to density.
}
