% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_Abundance_Viewer.R
\name{plotLineplotInd}
\alias{plotLineplotInd}
\title{Lineplot of cell population abundances displaying each individual}
\usage{
plotLineplotInd(
  CYTdata,
  marker = NULL,
  th = NULL,
  population = NULL,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  title = NULL,
  Yvalue = c("percentage", "absolute"),
  stats = TRUE,
  test.statistics = c("wilcox.test", "t.test"),
  paired = FALSE,
  corrMethod = "holm",
  BSL = "BSL2",
  meanMed = c("mean", "median"),
  palette = NULL
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{population}{a character vector containing the identifiers of the clusters/metaclusters to display in the plot. By default, all clusters or metaclusters are used.}

\item{level}{a character value containing the levels to display in the plot, either "clusters" (default) or "metaclusters"}

\item{samples}{a character vector containing the names of the biological samples to use. By default, all samples are used.}

\item{title}{a character value containing the title of the plot.}

\item{Yvalue}{a character value containing the variable to be studied. Possible values are percentage (default) or absolute.}

\item{stats}{a boolean value indicating whether statistical comparisons of each timepoint to the baseline should be performed.}

\item{test.statistics}{a character value providing the type of statistical test to use. Possible values are: 'wilcox.test' (default) or 't.test'}

\item{paired}{a boolean value indicating if a paired or unpaired comparison should be applied. Defaults to FALSE.}

\item{corrMethod}{a character value providing the type of correction for the p.value. Please refer to rstatix::wilcox_test or rstatix::t_test for details.}

\item{BSL}{a character value containing the name of the baseline sample against which the test will be performed for each timepoint. Defaults to "BSL2" (change ?)}

\item{meanMed}{a character value indicating whether to display the mean ('mean', default) or the median ('median') in the lineplot.}

\item{palette}{a character vector colors for the plots. if NULL (default), the CYTdata's colour palette will be used.}
}
\value{
a ggplot2 object
}
\description{
Similar to plotLineplot3, this function aims to visualize and compare the cell population abundances variation across timepoints for each biological condition using lineplot representations.
However, it does not display the mean and standard deviation, but the mean in bold and all other individuals with each an associated point shape. 

Several sets of clusters or metaclusters can theoretically be displayed, but it is recommended to use only one for plot readability.
The representation can be restricted to a specific set of samples.
Statistics can be computed for all comparisons.
}
