% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_Abundance_Viewer.R
\name{plotPiechart}
\alias{plotPiechart}
\title{Plot pie charts of population abundance}
\usage{
plotPiechart(
  CYTdata,
  samples = NULL,
  population = NULL,
  value = c("abundance", "count"),
  level = c("clusters", "metaclusters"),
  Xaxis = "Timepoint",
  Yaxis = "Individual",
  NFSvalues = NULL,
  pieSize = FALSE,
  byMeanSize = FALSE,
  colorBar = "#000000"
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{population}{a character vector containing the identifiers of the clusters/metaclusters to display in the plot. By default, all clusters or metaclusters are used.}

\item{value}{a character value containing the variable to be studied. Possible values are "abundance" (default, representing the relative abundance) or "count" (the cell count).
Please note that "count" is only relevant if pieSize is set to TRUE, and will make the pie vary in size depending on cell count.}

\item{level}{a character value containing the levels to display in the plot, either "clusters" (default) or "metaclusters"}

\item{Xaxis}{a string being the metadata used to plot pie charts on the horizontal line. By default, "Timepoint" metadata is used.}

\item{Yaxis}{a string being the metadata used to plot pie charts on the vertical line. By default, "Individual" metadata is used.}

\item{pieSize}{a boolean, controls pie size according to cell count. Bugged..}

\item{byMeanSize}{a boolean, controls pie size according to cell count. Bugged..}

\item{colorBar}{a character string indicating the color for the pie chart's borders. Defaults to black.}

\item{NFSValues}{a named numeric with blood count (numération formule sanguine) per sample. Optional. If provided, the abundance data will be multiplied by the blood count for pie size adjustement.}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the population composition of a CYTdata according to specified metadata (e.g., timepoint and individual), by plotting pie charts.
The pie charts are plotted on a 2D grid, with each axis showing one kind of metadata, allowing easy comparison of cell composition.  
Theoretically, the function should allow variations in the size of the pies to account for cell count. However, it is currently recommended to use value = "abundance" and to not set pieSize and byMeanSize to TRUE, as these parameters of the function are bugged.
}
