% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Clustering_Viewer.R
\name{plotPopulationCounts}
\alias{plotPopulationCounts}
\title{Plots the numbers of cells of each population (cluster or metacluster)}
\usage{
plotPopulationCounts(
  CYTdata,
  population = NULL,
  level = c("clusters", "metaclusters"),
  sort = TRUE,
  color.metadata = NULL
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{population}{a character vector containing the identifiers of the population to use. By default, all the population are used}

\item{level}{a character value indicating the type of population plotted. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{sort}{a boolean value indicating if population must be sorted by the number associated cluster}

\item{color.metadata}{a character value specifying the metadata used to color the barplot. By default, color.metadata is set to NULL (the barplot color is uniform)}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the number of cells associated to each population. The populations can be clusters or metaclusters

This representation displays the population in the X-axis and the total number of associated cells in the Y-axis.
}
