% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_Abundance_Viewer.R
\name{plotStackedArea}
\alias{plotStackedArea}
\title{Plot a stack area chart of cell population}
\usage{
plotStackedArea(
  CYTdata,
  population = NULL,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  observationMetadata = "Timepoint",
  Yvalue = c("ABrelative", "CCabsolute")
)
}
\arguments{
\item{CYTdata}{a CYTdata object.}

\item{population}{a character vector containing the names of the populations to use. By default, all clusters or metaclusters are used.}

\item{level}{a character value containing the levels to display in the plot, either "clusters" (default) or "metaclusters"}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{observationMetadata}{a character value containing the biological condition to use for the area chart, i.e., the column of the metadata dataframe. Defaults to "Timepoint".}

\item{Yvalue}{a character value specifying whether the total cell count for displayed metadata value should be normalized to 1 ("ABrelative", default) or if actual cell count should be displayed with varying plot height ("CCabsolute").}
}
\value{
a ggplot2 object
}
\description{
This function aims to display a stack area chart showing the abundance of different cell populations across specific metadata. 
It is particularly suited for visualization across timepoints.
}
