% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_Abundance_Viewer.R
\name{plotStackedBarplot}
\alias{plotStackedBarplot}
\title{Plot a barplot of cell population}
\usage{
plotStackedBarplot(
  CYTdata,
  population = NULL,
  level = c("clusters", "metaclusters"),
  samples = NULL,
  observationMetadata = "Timepoint",
  groupMetadata = NULL,
  Yvalue = c("ABrelative", "CCabsolute")
)
}
\arguments{
\item{CYTdata}{a CYTdata object.}

\item{population}{a character vector containing the names of the population to use.}

\item{level}{a character value containing the levels to display in the plot, either "clusters" (default) or "metaclusters"}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{observationMetadata}{a character value containing the biological condition to use for the barplots, i.e., the column of the metadata dataframe. Defaults to "Timepoint".}

\item{groupMetadata}{a character value containing the biological condition to use for further subdivision of the barplots. Optional, defaults to NULL.}

\item{Yvalue}{a character value specifying whether the total cell count for each barplot should be normalized to 1 ("ABrelative", default; i.e., each barplot at the same height) or if actual cell count should be displayed with varying barplot height ("CCabsolute")}
}
\value{
a ggplot2 object
}
\description{
This function aims to display a barplot showing the abundance of different cell populations across specific metadata.
}
