% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_DiffAbundanceAnalysis.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plots a volcano plot of a differential abundant analysis}
\usage{
plotVolcano(
  CYTdata,
  comparisonTitle,
  thPvalue = 0.05,
  thFoldchange = 2,
  displayAdjust = TRUE,
  signifOnly = TRUE
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{thPvalue}{a numeric value containing the p-value threshold to use. Defaults to 0.05}

\item{thFoldchange}{a numeric value containing the fold-change threshold to use for considering changes in population abundance between reference and comparison as significant. Defaults to 2.}

\item{displayAdjust}{a boolean. If TRUE, adjusted p-values will be used.}

\item{signifOnly}{a boolean. If TRUE, populations that do not show a significant change of abundance (in terms of pValue and foldchange) will be displayed in grey, with no name.}

\item{comparison}{a character value containing the name of the comparison to display}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the results of a differential abundant analysis using a Volcano plot
In such representation, each in dot corresponds to a phenotypic family (population; clusters or metaclusters). Dots are positioned in a two dimensional space where the X-axis represents the log2(fold-change) and the Y-axis represents the -log10 of the p-value.
One horizontal line shows the p-value threshold and two vertical lines show the fold-change threshold.
The populations are plotted and colored according to their color in the CYTdata.
}
