% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_Batch_Correction.R
\name{PVCA.batch}
\alias{PVCA.batch}
\title{Compute Principal Variance Component Analysis (PVCA) to determine the proportion of variability explained by selected effects (metadata condition)}
\usage{
PVCA.batch(CYTdata, batch.factors, threshold, markers = NULL)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{batch.factors}{a vector of factors that the mixed linear model will be fit on}

\item{threshold}{the percentile value of the minimum amount of the variabilities that the selected principal components need to explain}

\item{markers}{a character vector providing the cell markers to use. By default, all markers are used}
}
\value{
a list containing the PVCA features and a barplot (ggplot object)
}
\description{
This function is pvcaBatchAssess function from pvca package with some adaptations. This function assess the batch sources by fitting all "sources" as random effects
including two-way interaction terms in the Mixed Model(depends on lme4 package) to selected principal components, which were obtained from the original data correlation matrix.
}
