% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Clustering.R
\name{removeClustering}
\alias{removeClustering}
\title{Resets the 'Clustering' slot from a CYTdata object}
\usage{
removeClustering(CYTdata)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
The Clustering slot of a CYTdata object will be reset to an empty Clustering S4 object. 
The CYTdata object is validated (makeValid function) before and after removal.
}
