% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_DimReduction.R
\name{removeDimReduction}
\alias{removeDimReduction}
\title{Empty the 'DimReduction' slot in a CYTdata object}
\usage{
removeDimReduction(CYTdata)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Empty the 'DimReduction' slot in a CYTdata object
}
