% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_FCS_Markers_CYTdata.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{TO TEST Removes duplicated columns in a marker expression matrix}
\usage{
removeDuplicates(CYTdata, markers = NULL)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a character vector specifying a the name of the markers to test for duplication. If NULL, all markers will be tested.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to remove duplicated marker columns by checking whther marker expression is identicial across any pair of columns.
}
