% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Metaclustering.R
\name{removeMetaclustering}
\alias{removeMetaclustering}
\title{Resets the 'Metaclustering' slot from a CYTdata object}
\usage{
removeMetaclustering(CYTdata)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
The metaclustering slot of a CYTdata object will be reset to an empty Metaclustering S4 object. 
The CYTdata object is validated (makeValid function) before and after removal.
}
