% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_FCS_Markers_CYTdata.R
\name{renameMarkers}
\alias{renameMarkers}
\title{Renames markers within a CYTdata object}
\usage{
renameMarkers(
  CYTdata,
  from = NULL,
  to = NULL,
  removeConjugate = FALSE,
  sep = "_"
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{from}{a character vector providing the marker names to replace. By default, markers names from markers slot are replaced, in the order}

\item{to}{a character vector providing the new marker names to use}

\item{removeConjugate}{a boolean to indicate whether to remove the heavy metal name associated to the marker. Faster than specifying "to" manually, but specific.}

\item{sep}{a character specifying the separator of the metal and the marker name. Ignored if removeConjugate = F}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to rename cell markers stored within a CYTdata object.

This function is interesting to remove the names of the fluorochromes or metals recorded during the acquisition process.
}
