% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Population_Palette.R
\name{renamePopulations}
\alias{renamePopulations}
\title{Renames clusters or metaclusters within a CYTdata object.}
\usage{
renamePopulations(
  CYTdata,
  level = c("clusters", "metaclusters"),
  from = NULL,
  to,
  merge = FALSE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{level}{a character value indicating whether to rename clusters or metaclusters. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{from}{a character vector providing the cluster/metacluster names to replace. By default, all of them are used, in the level order.}

\item{to}{logical. If TRUE, clusters/metaclusters with duplicated names will be merged. If FALSE (default), duplicated names will result in an error.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to rename clusters or metaclusters stored within a CYTdata object. 
It can rename only a subset of the clusters/metaclusters, or all of them. 
If merge = TRUE, clusters/metaclusters which present the same name after renaming will be merged.
}
