% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Metadata_Samples.R
\name{renameSamples}
\alias{renameSamples}
\title{Renames samples within a CYTdata object}
\usage{
renameSamples(CYTdata, from = NULL, to, merge = FALSE)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{from}{a character vector providing the sample names to replace. By default, all sample names are replaced}

\item{to}{a character vector providing the new sample names to use}

\item{merge}{a boolean specifying whether to merge samples that have the same name after renaming. If set to FALSE, duplicated names will result in an error.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to rename samples stored within a CYTdata object. It can also merge samples based on their new names.
}
