% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Metaclustering.R
\name{rescale.exprs}
\alias{rescale.exprs}
\title{Internal - Rescales marker expression by quantile}
\usage{
rescale.exprs(exprs, rescaleBounds)
}
\arguments{
\item{exprs}{a vector containing one marker expression}

\item{rescaleBounds}{a numeric value providing the value of the first quantile and last quantile}
}
\value{
a numeric containing quantile rescale marker expressions
}
\description{
This function is used internally to rescale the marker expression by the quantile method.
Two quantiles are computed, and every expression value below the lower quantile or above the upper quantile is cropped to the respective quantile. Then, each expression value is rescaled from 0 to 1.
}
