% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_Batch_Correction.R
\name{run.cyCombine}
\alias{run.cyCombine}
\title{function for the cyCombine batch correction workflow.}
\usage{
run.cyCombine(
  CYTdata,
  batch.metadata,
  markers = NULL,
  norm_method = c("scale", "rank", "qnorm"),
  ties.method = c("average", "first", "last", "random", "max", "min"),
  seed = 92,
  xdim,
  ydim,
  rlen,
  covar.metadata = NULL,
  parametric = TRUE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{batch.metadata}{a character vector containing the names of metadata associated to batch acquisition (name of a CYTdata's metadata dataframe columns)}

\item{markers}{a character vector containing the names of biological markers to correct. By default, all Marker are corrected}

\item{norm_method}{a character being the normalization method. Should be either 'rank', 'scale' or 'qnorm'. Default: 'scale'. "rank" is recommended when combining data with heavy batch effects}

\item{ties.method}{a character being the method to handle ties, when using rank. Default: 'average'}

\item{seed}{a numeric being the seed to use when creating the SOM.}

\item{xdim}{a numeric being the x-dimension size of the SOM.}

\item{ydim}{a numeric being the y-dimension size of the SOM.}

\item{rlen}{a numeric being the number of times the data is presented to the SOM network. Consider a larger value, if results are not convincing (e.g. 100)}

\item{covar.metadata}{a character being the covariate ComBat should use. Must be the name of a specific metadata (name of a CYTdata's metadata dataframe columns)}

\item{parametric}{Default: TRUE. If TRUE, the parametric version of ComBat is used. If FALSE, the non-parametric version is used.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
Compute the batch correction on the data using the ComBat algorithm.
 Define a covariate, either as a character vector or name of tibble column.
 The covariate should preferable be the cell condition types but can be any column that infers heterogeneity in the data.
 The function assumes that the batch information is in the "batch" column and the data contains a "sample" column with sample information.
}
