% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_Batch_Correction.R
\name{run.CytoNorm}
\alias{run.CytoNorm}
\title{Function for the CytoNorm batch correction workflow.}
\usage{
run.CytoNorm(
  files.validation,
  labels.validation,
  files.training,
  labels.training,
  channels,
  transformList.training = NULL,
  transformList.validation = NULL,
  transformList.reverse.validation = NULL,
  FlowSOM.params = list(nCells = 1e+06, xdim = 15, ydim = 15, nClus = 30, scale = FALSE),
  normParams = list(nQ = 101, goal = "mean"),
  tmpDir = "./tmp",
  outputDir = "./CytoNorm",
  prefix = "CytoNorm_",
  seed = 1,
  clean = TRUE,
  verbose = FALSE,
  plot = FALSE,
  recompute = FALSE,
  write = TRUE,
  ...
)
}
\arguments{
\item{files.validation}{a character vector containing the paths of each data file which will be corrected}

\item{labels.validation}{a character vector containing the labels of batch acquisition variable for validation samples}

\item{files.training}{a character vector containing the paths of each data control file which will be used to train the model}

\item{labels.training}{labels.validation a character vector containing the labels of batch acquisition variable for validation samples}

\item{channels}{a character vector containing the names of channels to correct. By default, the common channels to the validation and training sets of files are selected}

\item{transformList.training}{Transformation list to pass to the flowCore to transform training data.
Default : arcsinh with cofactor 5 on all the common channels to the validation and training sets of files)}

\item{transformList.validation}{Transformation list to pass to the flowCore to transform validation data.
Default : same as transformList.training}

\item{transformList.reverse.validation}{Transformation list to pass to the flowCore to transform validation data after correction.
In an attempt to get raw corrected validation data. Default : reverse of transformList.validation transformation for the same channels}

\item{FlowSOM.params}{List with parameters to pass to the FlowSOM algorithm. Default = list(nCells = 1000000, xdim = 15, ydim = 15, nclus = 30, scale = FALSE).}

\item{tmpDir}{Directory to put the temporary files in. Default = "./tmp}

\item{outputDir}{Directory to put the temporary files in. Default = "."}

\item{prefix}{Prefix to put in front of the normalized file names. Default = "Norm_"}

\item{seed}{set.seed is called with this argument for reproducable results. Default = 1}

\item{clean}{Whether to remove the temporary files again at the end. Default = TRUE}

\item{verbose}{If TRUE, extra output is printed while running.}

\item{plot}{If TRUE, plots are saved to the current dir. Default = FALSE.}

\item{recompute}{If FALSE, will try to reuse previously saved FlowSOM model. If so, a warning message will be printed. Default = FALSE}

\item{write}{logical indicating whether the normalised samples should be written to new FCS files in a Normalized directory within outputDir, set to TRUE by default.}

\item{...}{Additional arguments to pass to read.FCS}
}
\value{
the paths of corrected FCS files
}
\description{
Compute the batch correction on the data using the CytoNorm algorithm. This algorithm requires control samples for each batch acquisition.
No CYTdata object is required, only the file path of conbtrol data that will train the model and dnon control data that will be corrected
For more infomation, see CytoNorm package : https://github.com/saeyslab/CytoNorm, https://rdrr.io/github/saeyslab/CytoNorm/man/
}
