% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Metaclustering.R
\name{run_Hierarchical_Metaclustering}
\alias{run_Hierarchical_Metaclustering}
\title{Performs hierarchical classification of the clusters (metaclustering)}
\usage{
run_Hierarchical_Metaclustering(
  CYTdata,
  markers = NULL,
  N_Metaclusters,
  N_NewClusters = NULL,
  criterion = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  typeMSI = c("median", "mean"),
  rescaleBounds = c(0.05, 0.95),
  seed = 42,
  checkOverwrite = TRUE
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a character vector providing the marker names to use. By default, all markers are used}

\item{N_Metaclusters}{a numeric value providing the number of metaclusters to regroup the clusters into}

\item{N_NewClusters}{a numeric value providing the number of new clusters to regroup the previous clusters into}

\item{criterion}{a string value providing the agglomeration method to be use.
Possible values are: 'ward.D', 'ward.D2', 'single', 'complete', 'average', 'mcquitty', 'median' or 'centroid'
Please refer to the function 'hclust' of the 'stats' package for details}

\item{typeMSI}{a string value providing the type of operation to compute on the marker expression of each cluster ("median" or "mean"; "median" by default)}

\item{seed}{a numeric value providing the random seed to use during stochastic operations}

\item{checkOverwrite}{a boolean value indicating whether to check if a Metaclustering has already been performed on the CYTdata object. If TRUE, it will check and ask for confirmation, if FALSE, any previous clustering will be overwritten by default.}

\item{rescale.bounds}{a numeric vector providing the two quantiles (within [0,1]) that will be used as bounds for rescaling the data. Defaults to c(0.05, 0.95)}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to identify metaclusters, which are groups of clusters having similar expressions for selected markers. 
The mean or median marker expressions is computed for each cluster, and used to compute hierarchical clustering using stats::hclust
Several clustering criteria are available, please refer to the function 'hclust' of the 'stats' package for details
Data will automatically be scaled by quantiles, by default 0.05 and 0.95

If N_NewClusters > 0, a second level clustering will be performed, where pre-existing clusters will be clustered into N_NewClusters new clusters, stored in the clustering slot of the CYTdata object (previous clusters will thus be overwritten)
}
