% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Clustering.R
\name{runClustering}
\alias{runClustering}
\title{Regroup cells with similar marker expressions using clustering techniques}
\usage{
runClustering(
  CYTdata,
  markers = NULL,
  type = c("SOM", "FlowSOM", "Phenograph", "Spade", "DBSCAN", "Kmeans", "Kmedoids",
    "CLARA", "PhenoGMM", "Mean_shift", "flowMeans"),
  scaledata = FALSE,
  checkOverwrite = TRUE,
  manifold = FALSE,
  seed = 42,
  ...
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a character vector providing the cell markers to use to generate the cluster data. By default, all markers are used}

\item{type}{a character value containing the type of clustering method to use. 
Possible values are: "SOM" (default), "FlowSOM", "Phenograph", "Spade", "DBSCAN", "Kmeans", "Kmedoids",  "CLARA" (, "PhenoGMM", "Mean_shift", "flowMeans").}

\item{scaledata}{boolean. If TRUE, centers (mean = 0) and scales (sd = 1) the expression of each marker}

\item{checkOverwrite}{a boolean value indicating whether to check if a Metaclustering has already been performed on the CYTdata object. 
If TRUE, it will check and ask for confirmation, if FALSE, any previous clustering will be overwritten by default.}

\item{manifold}{a boolean value. If TRUE, dimension reduction coordinates are used instead of cell marker expression to perform the clustering. 
Requires a dimension reduction to have been performed and stored in the CYTdata object. Defaults to FALSE.}

\item{seed}{a numeric value providing the random seed to use during stochastic operations}

\item{...}{additional arguments passed on to method from R package.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to regroup cells into clusters of cells with similar marker expression, based on a specific set of markers. 

Several algorithms are available, but as of 01/2026, only SOM has been thoroughly tested :
- Self-Organizing Maps (SOM), a neural network-based clustering technique (som method of the kohonen package)
- FlowSOM, self-organizing maps with the FlowSOM package (FlowSOM method). Also performs Metaclustering.
- Phenograph, a graph-based clustering technique designed for high-dimensional single cells, with cytofkit2 (Rphenograph method, requires Python)
- SPADE (Spanning-tree Progression Analysis of Density-normalized Events), see Qiu et al. 2011
- DBSCAN, a commonly used density-based clustering algorithm used for various kind of data (dbscan package)
- Kmeans, which clusters the cells in n clusters based on the distance between the cells and the centroid of each cluster (stats package)
- Kmedoids, more robust version of Kmeans (pam function of the cluster package)
- clara, similar to Kmedoids, but deals with large datasets better (clara function of the cluster package)
- "PhenoGMM", "Mean_shift", "flowMeans" were planned, but not implemented yet. 

The whole set of cell markers or specific cell markers can be used during the clustering process.

The cell clustering can be performed on marker expression (default), or on the coordinates obtained with a dimension reduction method.
}
