% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_DiffAbundanceAnalysis.R
\name{runDiffAbundanceAnalysis}
\alias{runDiffAbundanceAnalysis}
\title{Computes differential analysis statistics for cell clusters}
\usage{
runDiffAbundanceAnalysis(
  CYTdata,
  population = NULL,
  level = c("clusters", "metaclusters"),
  comparisonSpl,
  referenceSpl,
  comparisonTitle = "cmp",
  test.statistics = c("wilcox.test", "t.test"),
  p.adjust = c("holm", "none", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"),
  paired = FALSE,
  verbose = FALSE,
  NFSValues = NULL,
  checkOverwrite = TRUE
)
}
\arguments{
\item{CYTdata}{a CYTdata object}

\item{population}{a character vector containing the identifiers of the population to use. By default, all the population are used}

\item{level}{a character value indicating the type of population plotted. Possible values are: "clusters", "metaclusters". By default, 'clusters' are used.}

\item{comparisonSpl}{a character value providing the names of the samples to be compared to the reference samples}

\item{referenceSpl}{a character value providing the names of the reference samples}

\item{comparisonTitle}{a character value providing the name of comparison. Will be used for storage within the CYTdata object and for plotting with plotVolcano}

\item{test.statistics}{a character value providing the type of statistical test to use. Possible values are: 'wilcoxon.test' (default) or 't.test'}

\item{p.adjust}{a character value providing the type of p-value adjustment to use.
Possible values are: 'holm' (défault), 'none', 'hochberg', 'hommel', 'bonferroni', 'BH', 'BY', 'fdr'. See p.adjust function for details.}

\item{paired}{a boolean value indicating if a paired test should be performed. Defaults to FALSE.}

\item{verbose}{a boolean value. If TRUE, the function will print additional information.}

\item{NFSValues}{a named numeric with blood count (numération formule sanguine) per sample. Optional. If provided, the abundance data will be multiplied by the blood count before comparison.}

\item{checkOverwrite}{a boolean value. If TRUE, the function will check if a differential analysis with the same comparisonTitle has already been performed before overwriting.}
}
\value{
a S4 object of class 'CYTdata'
}
\description{
This function aims to identify differences of abundance between populations (i.e., clusters or metaclusters) in two sets of samples. 

The statistical test used for the comparisons can be defined by users, and may be paired or unpaired. Corrections can be computed for the p-value. 
For each population, the p-value, log2 fold-change and effect size relative to the reference comparisonSpl are computed.
}
