% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_DimReduction.R
\name{runDimReduction}
\alias{runDimReduction}
\title{Dimensionality Reduction of data using dimension reduction techniques}
\usage{
runDimReduction(
  CYTdata,
  markers = NULL,
  type = c("UMAP", "tSNE", "lvish", "PCA"),
  seed = 42,
  checkOverwrite = TRUE,
  ...
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{markers}{a character vector providing the cell markers to use to generate the reduced data}

\item{type}{a character value containing the type of RD method to use. Possible values are: "UMAP", "tSNE", "lvish", "PCA" (default = UMAP)}

\item{seed}{a numeric value providing the random seed to use during stochastic operations}

\item{checkOverwrite}{a boolean, if TRUE (default), will pause and display a warning if a DimReduction is already present in the CYTdata and must be overwritten.}

\item{...}{additional arguments passed on to method from R package.
For PCA, please refer to prcomp method from stats package : https://rdrr.io/r/stats/stats-package.html
For UMAP, please refer to umap method from uwot package : https://cran.r-project.org/web/packages/uwot/uwot.pdf
For t-SNE, please refer to Rtsne method from Rtsne package : https://cran.r-project.org/web/packages/Rtsne/Rtsne.pdf
For lvish, please refer to lvish method from uwot package : https://cran.r-project.org/web/packages/uwot/uwot.pdf}
}
\value{
a S4 object of class 'CYTdata', to which has been added a S4 object of class "DimReduction". 
DimReduction contains "coordinates", a n*2 matrix of the n cells coordinates on the 2D reduced space, 
and "optional_parameters", a list of "type", the name of the dimension reduction method used, "markers", the list of markers used, and "seed", the seed for stochastic operations.
}
\description{
This function aims to generate coordinates of data in an 2D-reduced space, stored in a CYTdata object.

The algorithm used available are :
- Principal component analysis (PCA), a linear dimension reduction technique.
- Uniform Manifold Approximation and Projection (UMAP), a non-linear dimension reduction technique
- t-distributed stochastic neighbor embedding (t-SNE), a non-linear dimension reduction technique
- LargeVis-like method (lvish), a non-linear dimension reduction technique
The whole set of cell markers or specific cell markers can be used during the dimensionality reduction process.
Note that for PCA, currently only the first two principal components are kept, with no way to check the percentage of variance explained by them. You may want to run the PCA yourself on CYTdata@matrix.expression[,markers] using prcomp while we integrate it better. 
Note also that t-SNE may require check_duplicates = FALSE.
}
