% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Population_Palette.R
\name{showPalettes}
\alias{showPalettes}
\title{Display the color palette associated with clusters or metaclusters as a plot}
\usage{
showPalettes(
  CYTdata,
  level = c("clusters", "metaclusters"),
  population = NULL,
  ncol = NULL,
  labelSize = 5
)
}
\arguments{
\item{CYTdata}{a S4 object of class 'CYTdata'}

\item{level}{a character value indicating whether to display clusters or metaclusters. Possible values are: "clusters" (default), "metaclusters".}

\item{population}{a character vector containing the identifiers of the clusters/metaclusters to display}

\item{ncol}{integer, number of columns in the graph. Defaults to one per 10 clusters/metaclusters}

\item{labelsize}{numeric, size of the population labels. Defaults to 5.}
}
\value{
nothing, but displays a plot.
}
\description{
Display the color palette associated with clusters or metaclusters as a plot
}
