% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_Abundance_Viewer.R
\name{statsLineplot}
\alias{statsLineplot}
\title{Stats function for lineplots (internal)}
\usage{
statsLineplot(
  dataMelted,
  statPairs,
  individuals,
  test.statistics = c("wilcox.test", "t.test"),
  paired = FALSE,
  corrMethod = "holm"
)
}
\arguments{
\item{dataMelted}{A melted dataframe as used in plotLineplot3}

\item{statPairs}{A list of character pairs indicating the timepoint pairs on which to compute tests}

\item{individuals}{A character vector containing the names of the individuals to consider.}

\item{test.statistics}{A character string indicating the type of statistics to compute : "wilcox.test" (default) or "t.test"}

\item{paired}{Logical. If TRUE, performs a pairwise test.}

\item{corrMethod}{A character string, the correction method for the p.value. Possible values are "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none", see p.adjust for details.}
}
\value{
a dataframe containing the number/percentage of cells per sample for the given pop, along with collated metadata
}
\description{
Computes pairwise abundance comparison between timepoints and a reference timepoint (usually the baseline) 
Currently (2026-03-02) bugged if less than three pairs to compute.
}
